// examples/better-auth/src/generated/prisma/models/Idea.ts
export type IdeaCountOutputTypeSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  newsPosts?: boolean | IdeaCountOutputTypeCountNewsPostsArgs
}

/**
 * IdeaCountOutputType without action
 */
export type IdeaCountOutputTypeDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the IdeaCountOutputType
   */
  select?: Prisma.IdeaCountOutputTypeSelect<ExtArgs> | null
}

/**
 * IdeaCountOutputType without action
 */
export type IdeaCountOutputTypeCountNewsPostsArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  where?: Prisma.IdeaNewsPostWhereInput
}


export type IdeaSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  title?: boolean
  content?: boolean
  description?: boolean
  image?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  userId?: boolean
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
  newsPosts?: boolean | Prisma.Idea$newsPostsArgs<ExtArgs>
  _count?: boolean | Prisma.IdeaCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["idea"]>
// ...
export type Idea$newsPostsArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the IdeaNewsPost
   */
  select?: Prisma.IdeaNewsPostSelect<ExtArgs> | null
  /**
   * Omit specific fields from the IdeaNewsPost
   */
  omit?: Prisma.IdeaNewsPostOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.IdeaNewsPostInclude<ExtArgs> | null
  where?: Prisma.IdeaNewsPostWhereInput
  orderBy?: Prisma.IdeaNewsPostOrderByWithRelationInput | Prisma.IdeaNewsPostOrderByWithRelationInput[]
  cursor?: Prisma.IdeaNewsPostWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.IdeaNewsPostScalarFieldEnum | Prisma.IdeaNewsPostScalarFieldEnum[]
}